requirements.txt
python-telegram-bot==13.15
requests==2.31.0
python-dotenv==1.0.0
SQLAlchemy==2.0.20
uuid==1.30